refactor: 패키지 구조를 domain/random 중심으로 이동
