+++ src/main/java/racingcar/random/RandomNumber.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.random;
 
 public interface RandomNumber {
     int generate();
