+++ src/main/java/racingcar/random/SimpleRandomNumber.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.random;
 
 import java.util.Random;
